# api/openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: Imagine API
  version: 0.1.0
  description: OpenAPI specification for the Imagine API (chi-based)
servers:
  - url: http://localhost:7770
    description: Local API
paths:
...
  /collections:
    get:
      summary: List collections
      operationId: listCollections
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Collections page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionListResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create a collection
      operationId: createCollection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /collections/{uid}:
    get:
      summary: Get collection detail
      operationId: getCollection
      parameters:
        - in: path
          name: uid
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Collection detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionDetailResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      summary: Update collection
      operationId: updateCollection
      parameters:
        - in: path
          name: uid
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionUpdate"
      responses:
        "200":
          description: Collection updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Delete a collection
      operationId: deleteCollection
      parameters:
        - in: path
          name: uid
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content (collection deleted)
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /collections/{uid}/images:
    get:
      summary: List images in a collection
      operationId: listCollectionImages
      parameters:
        - in: path
          name: uid
          required: true
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Images page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImagesPage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      summary: Add images to a collection
      operationId: addCollectionImages
      parameters:
        - in: path
          name: uid
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                uids:
                  type: array
                  items:
                    type: string
              required: [uids]
      responses:
        "200":
          description: Images added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddImagesResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddImagesResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddImagesResponse"
    delete:
      summary: Remove images from a collection
      operationId: deleteCollectionImages
      parameters:
        - in: path
          name: uid
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                uids:
                  type: array
                  items:
                    type: string
              required: [uids]
      responses:
        "200":
          description: Images removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteImagesResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteImagesResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteImagesResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    AddImagesResponse:
      type: object
      properties:
        added:
          type: boolean
        error:
          type: string
      required: [added]
    DeleteImagesResponse:
      type: object
      properties:
        deleted:
          type: boolean
        error:
          type: string
      required: [deleted]
...
    ImageEXIF:
      type: object
      properties:
        exif_version: { type: string }
        make: { type: string }
        model: { type: string }
        date_time: { type: string }
        date_time_original: { type: string }
        iso: { type: string }
        focal_length: { type: string }
        exposure_time: { type: string }
        aperture: { type: string }
        flash: { type: string }
        white_balance: { type: string }
        lens_model: { type: string }
        modify_date: { type: string }
        rating: { type: string }
        orientation: { type: string }
        resolution: { type: string }
        software: { type: string }
        longitude: { type: string }
        latitude: { type: string }
    ImagePaths:
      type: object
      properties:
        original: { type: string }
        thumbnail: { type: string }
        preview: { type: string }
        raw: { type: string }
      required: [original, thumbnail, preview]
    ImageMetadata:
      type: object
      properties:
        file_name: { type: string }
        file_size: { type: integer, format: int64 }
        original_file_name: { type: string }
        file_type: { type: string }
        keywords:
          type: array
          items: { type: string }
        color_space: { type: string }
        file_modified_at: { type: string, format: date-time }
        file_created_at: { type: string, format: date-time }
        thumbhash: { type: string }
        label: { type: string }
        checksum: { type: string }
      required:
        [
          file_name,
          file_type,
          color_space,
          file_modified_at,
          file_created_at,
          checksum,
        ]
    Image:
      type: object
      properties:
        uid: { type: string }
        name: { type: string }
        uploaded_by:
          $ref: "#/components/schemas/User"
        description: { type: string }
        exif:
          $ref: "#/components/schemas/ImageEXIF"
        private: { type: boolean }
        width: { type: integer, format: int32 }
        height: { type: integer, format: int32 }
        processed: { type: boolean }
        image_metadata:
          $ref: "#/components/schemas/ImageMetadata"
        image_paths:
          $ref: "#/components/schemas/ImagePaths"
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required:
        [
          uid,
          name,
          private,
          width,
          height,
          processed,
          created_at,
          updated_at,
          image_paths,
        ]
    CollectionImage:
      type: object
      properties:
        uid: { type: string }
        added_at: { type: string, format: date-time }
        added_by:
          $ref: "#/components/schemas/User"
      required: [uid, added_at]
    Collection:
      type: object
      properties:
        uid: { type: string }
        name: { type: string }
        image_count: { type: integer }
        private: { type: boolean, nullable: true }
        images:
          type: array
          items:
            $ref: "#/components/schemas/CollectionImage"
        created_by:
          $ref: "#/components/schemas/User"
        description: { type: string }
        thumbnail:
          $ref: "#/components/schemas/Image"
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [uid, name, image_count, created_at, updated_at]
    CollectionCreate:
      type: object
      properties:
        name: { type: string }
        private: { type: boolean, nullable: true }
        description: { type: string }
      required: [name]
    CollectionUpdate:
      type: object
      properties:
        name: { type: string }
        thumbnailUID: { type: string }
        description: { type: string }
        private: { type: boolean }
        ownerUID: { type: string }
    ImagesResponse:
      type: object
      properties:
        added_at: { type: string, format: date-time }
        added_by:
          $ref: "#/components/schemas/User"
        image:
          $ref: "#/components/schemas/Image"
      required: [added_at, image]
    ImagesPage:
      type: object
      properties:
        href: { type: string }
        prev: { type: string }
        next: { type: string }
        limit: { type: integer }
        offset: { type: integer }
        count: { type: integer }
        items:
          type: array
          items:
            $ref: "#/components/schemas/ImagesResponse"
      required: [limit, offset, items]
    CollectionListResponse:
      type: object
      properties:
        href: { type: string }
        prev: { type: string }
        next: { type: string }
        limit: { type: integer }
        offset: { type: integer }
        count: { type: integer }
        items:
          type: array
          items:
            $ref: "#/components/schemas/Collection"
      required: [limit, offset, items]
    CollectionDetailResponse:
      type: object
      properties:
        uid: { type: string }
        name: { type: string }
        image_count: { type: integer }
        private: { type: boolean, nullable: true }
        images:
          $ref: "#/components/schemas/ImagesPage"
        created_by:
          $ref: "#/components/schemas/User"
        description: { type: string }
        thumbnail:
          $ref: "#/components/schemas/Image"
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [uid, name, images, created_at, updated_at]
    User:
      type: object
      properties:
        uid: { type: string }
        first_name: { type: string }
        last_name: { type: string }
        username: { type: string }
        email: { type: string }
        role:
          type: string
          enum: [user, admin, superadmin, guest]
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required:
        [
          uid,
          first_name,
          last_name,
          username,
          email,
          role,
          created_at,
          updated_at,
        ]
...